Make FileStore thread safe
